Implemented PWM output for Peltier
Experimental fix for retrofitted oversized Peltier device that is now
exceeding the power dissipation of the stock heat sink, which is causing
the temperature inside the cooler to run warmer than it originally did.
Testing at 50% PWM to see if reducing the current flow through the
Peltier will bring the system back to an optimum cooling state.